requirements.txt
streamlit==1.24.0
pandas==2.0.2
plotly==5.15.0
numpy==1.24.3
psycopg2-binary==2.9.9

# UI компоненты
altair==5.0.1